Se realiza avance en menu desplegable

+++ src/app/pages/private/products/get-products-admin.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
 import { ProductService } from '../../../services/product-service';
 import { CurrencyPipe } from '@angular/common';
+import Swal from 'sweetalert2';
 @Component({
   selector: 'app-get-products-admin',
   imports: [CurrencyPipe],
@@ -9,6 +10,7 @@ import { CurrencyPipe } from '@angular/common';
 })
 export class GetProductsAdmin {
   products: any = [];
+  filterProducts: any[] = [];
 
   constructor(
     private productService: ProductService
@@ -29,19 +31,44 @@ export class GetProductsAdmin {
   }
 
   onDelete(id: string) {
-    console.log(id)
-    this.productService.deleteProducts(id).subscribe({
-      next: (data) => {
-        console.log(data);
-        this.ngOnInit()
-      },
-      error: (error) => {
-        console.error(error);
-      },
-      complete: () => { }
+
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: "El producto será eliminado.",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sí, ¡eliminalo!',
+      cancelButtonText: 'Cancelar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        console.log(id)
+        this.productService.deleteProducts(id).subscribe({
+          next: (data) => {
+            console.log(data);
+            this.ngOnInit()
+          },
+          error: (error) => {
+            console.error(error);
+          },
+          complete: () => { }
+        });
+      }
     });
   }
 
-  
+  // buscarProductos(filtro: string) {
+  //   this.productService.filterProducts().subscribe({
+  //     next: (res: any[]) => {
+  //       this.filterProducts = res; 
+  //       console.log("Resultado filtrado:", res);
+  //     },
+  //     error: (err) => {
+  //       console.error("Error al filtrar:", err);
+  //     }
+  //   });
+  // }
+
 }
